+++ Modules.php
@@ -27,16 +27,6 @@ if ( ! isset( $_REQUEST['modfunc'] ) )
 
 $_ROSARIO['page'] = 'modules';
 
-// Save $_REQUEST vars in session: used to recreate $_REQUEST in Bottom.php.
-if ( ! isset( $_REQUEST['_ROSARIO_PDF'] )
-	&& empty( $_REQUEST['LO_save'] )
-	&& ( mb_strpos( $modname, 'misc/' ) === false
-		|| $modname === 'misc/Portal.php' )
-	&& $modname !== 'Reports/SavedReports.php' )
-{
-	$_SESSION['_REQUEST_vars'] = $_REQUEST;
-}
-
 // Set Popup window detection.
 isPopup( $modname, $_REQUEST['modfunc'] );
 
@@ -124,6 +114,20 @@ elseif ( User( 'USERNAME' ) )
 	require_once 'ProgramFunctions/HackingLog.fnc.php';
 
 	HackingLog();
+}
+
+/**
+ * Save $_REQUEST vars in session: used to recreate $_REQUEST in Bottom.php
+ * Note: Code duplicated inside ErrorMessage() for fatal errors
+ *
+ * @since 11.5 Copy $_REQUEST to $_SESSION['_REQUEST_vars'] last
+ */
+if ( ! isset( $_REQUEST['_ROSARIO_PDF'] )
+	&& empty( $_REQUEST['LO_save'] )
+	&& ( mb_strpos( $modname, 'misc/' ) === false
+		|| $modname === 'misc/Portal.php' ) )
+{
+	$_SESSION['_REQUEST_vars'] = $_REQUEST;
 }
 
 // Output Footer HTML.