+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 11.5
 - Moodle plugin: Add REST API protocol in plugins/Moodle/, rosariosis.sql & rosariosis_mysql.sql
 - Add MoodleAPICall() & MoodleRESTCall() functions in plugins/Moodle/client.php
 - Send Course description HTML to Moodle in plugins/Moodle/Scheduling/Courses.php
+- Copy `$_REQUEST` to `$_SESSION['_REQUEST_vars']` last in Modules.php, ErrorMessage.fnc.php & PreparePHP_SELF.fnc.php
 
 Changes in 11.4.2
 -----------------