+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 11.5
 - Add MoodleAPICall() & MoodleRESTCall() functions in plugins/Moodle/client.php
 - Send Course description HTML to Moodle in plugins/Moodle/Scheduling/Courses.php
 - Copy `$_REQUEST` to `$_SESSION['_REQUEST_vars']` last in Modules.php, ErrorMessage.fnc.php & PreparePHP_SELF.fnc.php
+- No error message if error code 4 (no file was attached) in FileUpload.fnc.php
 
 Changes in 11.4.2
 -----------------