+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 11.5
 - Send Course description HTML to Moodle in plugins/Moodle/Scheduling/Courses.php
 - Copy `$_REQUEST` to `$_SESSION['_REQUEST_vars']` last in Modules.php, ErrorMessage.fnc.php & PreparePHP_SELF.fnc.php
 - No error message if error code 4 (no file was attached) in FileUpload.fnc.php
+- DBInsertSQL() & DBUpdateSQL() better check empty values ('', false, null) in DBUpsert.php
 
 Changes in 11.4.2
 -----------------